5. Metrics calculation/requirements.txt
pyyaml
pandas
openpyxl
openai
pydantic
tqdm
